+++ src/Rebuy/Amqp/Consumer/ConsumerManager.php
@@ -6,8 +6,8 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use PhpAmqpLib\Channel\AMQPChannel;
 use PhpAmqpLib\Message\AMQPMessage;
-use Rebuy\Amqp\Consumer\Annotation\ConsumerContainer;
-use Rebuy\Amqp\Consumer\Annotation\Parser;
+use Rebuy\Amqp\Consumer\Attribute\ConsumerContainer;
+use Rebuy\Amqp\Consumer\Attribute\Parser;
 use Rebuy\Amqp\Consumer\Exception\ConsumerContainerException;
 use Rebuy\Amqp\Consumer\Exception\ConsumerException;
 use Rebuy\Amqp\Consumer\Handler\ErrorHandlerInterface;
