+++ tests/Rebuy/Tests/Amqp/Consumer/ConsumerManagerTest.php
@@ -8,8 +8,8 @@ use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
-use Rebuy\Amqp\Consumer\Annotation\ConsumerContainer;
-use Rebuy\Amqp\Consumer\Annotation\Parser;
+use Rebuy\Amqp\Consumer\Attribute\ConsumerContainer;
+use Rebuy\Amqp\Consumer\Attribute\Parser;
 use Rebuy\Amqp\Consumer\ConsumerManager;
 use Rebuy\Amqp\Consumer\Exception\ConsumerException;
 use Rebuy\Amqp\Consumer\Serializer\Serializer;
