+++ tests/Rebuy/Tests/Amqp/Consumer/Stubs/ConsumerWithAttributes.php
@@ -4,7 +4,7 @@ namespace Rebuy\Tests\Amqp\Consumer\Stubs;
 
 class ConsumerWithAttributes
 {
-    #[\Rebuy\Amqp\Consumer\Annotation\Consumer(name: 'consume-with-attributes', prefetchCount: 100)]
+    #[\Rebuy\Amqp\Consumer\Attribute\Consumer(name: 'consume-with-attributes', prefetchCount: 100)]
     public function consume(Message $message): void
     {
     }
